+++ app/Importers/PnpTrienaleImporter.php
@@ -23,8 +23,6 @@ class PnpTrienaleImporter extends AbstractImporter
         'measurement:cs' => 'Rozměry:',
         'related_work:sk' => 'Ze souboru:',
         'related_work:cs' => 'Ze souboru:',
-        'description:sk' => 'Text od porotců:',
-        'description:cs' => 'Text od porotců:',
     ];
 
     protected $defaults = [
@@ -161,6 +159,15 @@ class PnpTrienaleImporter extends AbstractImporter
         }
 
         return implode(';', $techniques);
+    }
+
+    public function hydrateDescription(array $record, $locale)
+    {
+        if ($locale !== 'cs' && $locale !== 'sk') {
+            return null;
+        }
+
+        return htmlentities($record['Text od porotců:']);
     }
 
     public function hydrateAdditionals(array $record, $locale)